+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## 0.8.0
 
-Feat: added migrater module.
+- Feat: added migrater module.
+
+- Feat: allow to add lambda package by installer.